modified response-request

+++ checkout/src/main/java/company/tap/checkout/internal/interfaces/OnCurrencyChangedActionListener.kt
@@ -1,5 +1,7 @@
 package company.tap.checkout.internal.interfaces
 
+import java.math.BigDecimal
+
 
 /**
  * Created by AhlaamK on 12/9/20.
@@ -8,5 +10,5 @@ Copyright (c) 2020    Tap Payments.
 All rights reserved.
  **/
 interface OnCurrencyChangedActionListener {
-    fun onCurrencyClicked(currencySelected: String, currencyRate: Double)
+    fun onCurrencyClicked(currencySelected: String, currencyRate: BigDecimal)
 }
